requirements.txt
streamlit>=1.0
pandas>=1.3
numpy>=1.21
tensorflow>=2.8
plotly>=5.0
matplotlib>=3.4
scikit-learn>=1.0
Pillow>=9.0
pytesseract>=0.3
spacy>=3.0
shap>=0.41
